ds/requirements.txt
numpy
scipy
pandas
pyproj
matplotlib
jupyter
ipython
nuts-finder
setuptools
git+ssh://git@github.com/nestauk/data_getters.git
ratelim
geopandas
shapely
rtree

# local package
-e .

# external requirements
dvc
jq
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
